rough refactor of stitching
